rename old bulk publish tool
