+++ libs/blocks/bulk-publish-v1/bulk-publish.js
